add env vars to jack config script

+++ src/jack_config_script/save_config.sh
@@ -2,6 +2,13 @@
 
 cd "$(dirname "`readlink -f "$(realpath "$0")"`")"
 source shared.sh || exit 0
-contents=$(get_current_parameters) || exit 0
+contents=$(get_current_parameters)
+
+if $RAY_FAIL_IF_JACK_DIFF && [ -f "$session_jack_file" ];then
+    wanted_parameters=$(cat "$session_jack_file") || exit 3
+    make_diff_parameters
+    [ -n "$diff_parameters" ] && exit 29
+fi
+
 echo "$contents" > "$session_jack_file"
 exit 0
